separated types file and generateQR return DataURL no
