+++ lib/index.ts
@@ -1,6 +1,6 @@
-import { validateTx } from "./functions/validatetx"
-import { onPayment } from "./functions/onpayment"
-import { generatePaymentLink } from "./functions/generatepaymentlink";
-import { generatePaymentQR } from "./functions/generatepaymentqr";
+import { validateTx } from "./utils/validatetx"
+import { onPayment } from "./utils/onpayment"
+import { generatePaymentLink } from "./utils/generatepaymentlink";
+import { generatePaymentQR } from "./utils/generatepaymentqr";
 
 export {validateTx, onPayment, generatePaymentLink, generatePaymentQR};
