+++ lib/utils/generatepaymentlink.ts
@@ -1,24 +1,4 @@
-export interface VendorMetadata {
-    callbackUrl: URL
-  
-    invoiceId: string
-  
-    website: URL
-  
-    vendorName: string
-}
-
-export interface PaymentDetails{
-    recipient: string
-
-    amount: number
-
-    reason: string
-
-    category: string
-
-    vendorData?: VendorMetadata
-}
+import { PaymentDetails } from "../types";
 
 
 export function generatePaymentLink(PaymentDetails:PaymentDetails) : string {
