+++ lib/utils/validatetx.ts
@@ -1,0 +1,4 @@
+import { TxData } from "../types";
+export function validateTx(TxData:TxData):boolean{
+    return true;
+}
